Optimize oppacity and refactor code with pyright

+++ threeBodyProblem/simulation_params.py
@@ -5,7 +5,7 @@ from threeBodyProblem.constants import PHYSICS_CONSTANTS
 
 @dataclass
 class SimulationParams:
-    alpha: float = 255,
-    body_distance: float = PHYSICS_CONSTANTS.DEFAULT_BODY_DISTANCE,
-    mass: float = PHYSICS_CONSTANTS.DEFAULT_BODY_MASS,
+    alpha: int = 255
+    body_distance: float = PHYSICS_CONSTANTS.DEFAULT_BODY_DISTANCE
+    mass: float = PHYSICS_CONSTANTS.DEFAULT_BODY_MASS
     g_constant: float = PHYSICS_CONSTANTS.GRAVITATIONAL_CONSTANT
